Linter fixes

+++ saturnin/src/video/vdp1.cpp
@@ -97,7 +97,7 @@ void Vdp1::onVblankIn() {
 
 auto Vdp1::vdp1Parts() const -> std::vector<Vdp1Part> { return vdp1_parts_; }
 
-auto Vdp1::vdp1Parts(const u32 priority) const -> std::vector<Vdp1Part> {
+auto Vdp1::vdp1Parts(const u8 priority) const -> std::vector<Vdp1Part> {
     auto parts = std::vector<video::Vdp1Part>{};
     std::ranges::copy_if(vdp1_parts_, std::back_inserter(parts), [priority](const Vdp1Part& part) {
         return part.common_vdp_data_.priority == priority;
